requirements.txt
streamlit
fastapi
uvicorn
langchain
pydantic
langchain-openai
python-dotenv
langchain-community
langchain-tavily
langgraph
langchain-groq
langchain-huggingface
sentence-transformers
langchain-chroma

